[core] Enable local glyph generation using TinySDF.  - Platform-specific LocalGlyphRasterizer is responsible for deciding which glyphs to rasterize locally and for implementing the rasterization.  - Default platform implementation doesn't locally generate any glyphs -> no behavior change  - Unit test uses StubLocalGlyphRasterizer, which returns a single fixed bitmap for all CJK glyphs  - Rename glyph_loader.test to glyph_manager.test
