+++ cmake/test-files.cmake
@@ -113,7 +113,7 @@ set(MBGL_TEST_FILES
 
     # text
     test/text/cross_tile_symbol_index.test.cpp
-    test/text/glyph_loader.test.cpp
+    test/text/glyph_manager.test.cpp
     test/text/glyph_pbf.test.cpp
     test/text/quads.test.cpp
